Implement basic function of Cold Redundancy
Implement basic function of Cold Redundancy include rotation, config,
check and reRanking.

Tested:
Change rotationPeriod to a smaller value such as 60 seconds. Insert
two PSU on system, and check PSU rank order register 0xd0. After 60s
The value of rank order in PSU1 will change to 2, and the value of
rank order in PSU2 will change to 1.
If removed AC Cable on one PSU. The rank order in this PSU will be
always 0, and the rank order in the other PSU will be always 1.

Change-Id: Idc14f783ec83db005ecf9fc53332431ba6a6daee
Signed-off-by: Cheng C Yang <cheng.c.yang@intel.com>